F-1698 F-1699 F-2094 F-2097 F-3222 F-3465 F-3900 F-4130 F-4599 F-5611 F-6285: fix logic and conditional bugs

+++ pkcs7/envelopedData-ktri-stream.c
@@ -101,11 +101,11 @@ static int load_certs(byte* cert, word32* certSz, byte* key, word32* keySz)
 }
 
 
-static size_t envelopedData_encrypt(byte* cert, word32 certSz, byte* key,
+static int envelopedData_encrypt(byte* cert, word32 certSz, byte* key,
                                  word32 keySz, byte* out, word32 outSz,
                                  word32 contentSz, byte useStreamMode)
 {
-    size_t ret;
+    int ret;
     PKCS7* pkcs7;
 
     pkcs7 = wc_PKCS7_New(NULL, INVALID_DEVID);
@@ -125,7 +125,7 @@ static size_t envelopedData_encrypt(byte* cert, word32 certSz, byte* key,
     /* add recipient using RSA certificate (KTRI type) */
     ret = wc_PKCS7_AddRecipient_KTRI(pkcs7, cert, certSz, 0);
     if (ret < 0) {
-        printf("wc_PKCS7_AddRecipient_KTRI() failed, ret = %zu\n", ret);
+        printf("wc_PKCS7_AddRecipient_KTRI() failed, ret = %d\n", ret);
         wc_PKCS7_Free(pkcs7);
         return -1;
     }
@@ -133,7 +133,7 @@ static size_t envelopedData_encrypt(byte* cert, word32 certSz, byte* key,
     /* encode envelopedData, returns size */
     ret = wc_PKCS7_EncodeEnvelopedData(pkcs7, out, outSz);
     if (ret <= 0) {
-        printf("ERROR: wc_PKCS7_EncodeEnvelopedData() failed, ret = %zu\n", ret);
+        printf("ERROR: wc_PKCS7_EncodeEnvelopedData() failed, ret = %d\n", ret);
         wc_PKCS7_Free(pkcs7);
         return -1;
 
@@ -147,11 +147,11 @@ static size_t envelopedData_encrypt(byte* cert, word32 certSz, byte* key,
     return ret;
 }
 
-static size_t envelopedData_decrypt(byte* in, word32 inSz, byte* cert,
+static int envelopedData_decrypt(byte* in, word32 inSz, byte* cert,
                                 word32 certSz, byte* key, word32 keySz,
                                 byte* out, word32 outSz)
 {
-    size_t ret;
+    int ret;
     PKCS7* pkcs7;
 
     pkcs7 = wc_PKCS7_New(NULL, INVALID_DEVID);
@@ -175,7 +175,7 @@ static size_t envelopedData_decrypt(byte* in, word32 inSz, byte* cert,
     /* decode envelopedData, returns size */
     ret = wc_PKCS7_DecodeEnvelopedData(pkcs7, in, inSz, out, outSz);
     if (ret <= 0) {
-        printf("Failed to decode EnvelopedData bundle (%s), error %zu\n",
+        printf("Failed to decode EnvelopedData bundle (%s), error %d\n",
                 encodedFileKTRI, ret);
         wc_PKCS7_Free(pkcs7);
         return -1;
@@ -197,7 +197,7 @@ static size_t envelopedData_decrypt(byte* in, word32 inSz, byte* cert,
 int main(int argc, char** argv)
 {
     int ret = 0;
-    size_t encryptedSz = 0, decryptedSz;
+    int encryptedSz = 0, decryptedSz;
     word32 certSz, keySz, contentSz = 0;
 
     byte cert[DEFAULT_EXAMPLE_BUFFER_SIZE];
@@ -279,7 +279,7 @@ int main(int argc, char** argv)
             printf("error reading file %s\n", encodedFileKTRI);
             ret = -1;
         }
-        printf("Read %ld bytes for encrypted file found\n", encryptedSz);
+        printf("Read %d bytes for encrypted file found\n", encryptedSz);
     }
 
     if (ret == 0) {
